lib: optimized int_sqrt for small ints
Modified the int_sqrt function to check for smaller ints
to decrease cpu idle time for smaller ints. This patch is
based on the fact that most ints being called are smaller
than 64 bit ints, so the function now checks for them first,
and if they are found then it reduces the runtime of the
sqrt algorithm. There is a small hit in efficiency in 64
bit ints, but since they are called at a much lower frequency
than smaller ints.